+++ src/components/dialogs/inputs/input-hooks.js
@@ -52,10 +52,6 @@ import {
     genHelperPreviousValue,
 } from './hooks-helpers';
 import { useAutocompleteField } from './use-autocomplete-field';
-import Grid from '@mui/material/Grid';
-import { Box } from '@mui/system';
-import DeleteIcon from '@mui/icons-material/Delete';
-import AddIcon from '@mui/icons-material/ControlPoint';
 import RegulatingTerminalEdition, {
     makeRefreshRegulatingTerminalSectionsCallback,
 } from '../regulating-terminal-edition';
@@ -565,137 +561,6 @@ export const useRegulatingTerminalValue = ({
 
     return [regulatingTerminal, render];
 };
-
-// TODO CHARLY remove below
-export const useTableValues = ({
-    id,
-    tableHeadersIds,
-    Field,
-    inputForm,
-    defaultValues,
-    isReactiveCapabilityCurveOn,
-    disabled = false,
-}) => {
-    const [values, setValues] = useState([]);
-    const classes = useStyles();
-
-    const handleAddValue = useCallback(() => {
-        setValues((oldValues) => [...oldValues, {}]);
-    }, []);
-
-    const checkValues = useCallback(() => {
-        if (defaultValues !== undefined && defaultValues.length !== 0) {
-            setValues([...defaultValues]);
-        } else {
-            setValues([]);
-            handleAddValue();
-        }
-    }, [defaultValues, handleAddValue]);
-
-    useEffect(() => {
-        checkValues();
-    }, [checkValues]);
-
-    const handleDeleteItem = useCallback(
-        (index) => {
-            setValues((oldValues) => {
-                let newValues = [...oldValues];
-                newValues.splice(index, 1);
-                return newValues;
-            });
-            inputForm.reset();
-        },
-        [inputForm]
-    );
-
-    const handleSetValue = useCallback((index, newValue) => {
-        setValues((oldValues) => {
-            let newValues = [...oldValues];
-            newValues[index] = newValue;
-            return newValues;
-        });
-    }, []);
-
-    useEffect(() => {
-        // TODO CHARLY Surveiller cette fonction
-        if (!isReactiveCapabilityCurveOn) {
-            //TODO When isReactiveCapabilityCurveOn is false, the reactive capability curve component does not change
-            // the validation of its values and they still required.
-            // we update the validations of reactive capability curve values so they are not required any more.
-            // is there a better way to do it ?
-            function validate() {
-                return !isReactiveCapabilityCurveOn;
-            }
-
-            values.forEach((value, index) => {
-                inputForm.addValidation('P' + index, validate);
-                inputForm.addValidation('QmaxP' + index, validate);
-                inputForm.addValidation('QminP' + index, validate);
-            });
-        }
-    }, [inputForm, values, isReactiveCapabilityCurveOn]);
-
-    const field = useMemo(() => {
-        return (
-            <Grid item container spacing={2}>
-                {tableHeadersIds.map((header) => (
-                    <Grid key={header} item xs={3}>
-                        <FormattedMessage id={header} />
-                    </Grid>
-                ))}
-                <Box sx={{ width: '100%' }} />
-                {values.map((value, idx) => (
-                    <Grid key={id + idx} container spacing={3} item>
-                        <Field
-                            defaultValue={value}
-                            onChange={handleSetValue}
-                            index={idx}
-                            inputForm={inputForm}
-                            isFieldRequired={isReactiveCapabilityCurveOn}
-                            disabled={disabled}
-                        />
-                        <Grid item xs={1}>
-                            <IconButton
-                                className={classes.icon}
-                                key={id + idx}
-                                onClick={() => handleDeleteItem(idx)}
-                                disabled={disabled || idx === 0}
-                            >
-                                <DeleteIcon />
-                            </IconButton>
-                        </Grid>
-                        {idx === values.length - 1 && (
-                            <Grid item xs={1}>
-                                <IconButton
-                                    className={classes.icon}
-                                    key={id + idx}
-                                    onClick={() => handleAddValue()}
-                                    disabled={disabled}
-                                >
-                                    <AddIcon />
-                                </IconButton>
-                            </Grid>
-                        )}
-                    </Grid>
-                ))}
-            </Grid>
-        );
-    }, [
-        values,
-        id,
-        classes.icon,
-        handleAddValue,
-        handleDeleteItem,
-        handleSetValue,
-        inputForm,
-        tableHeadersIds,
-        isReactiveCapabilityCurveOn,
-        disabled,
-    ]);
-
-    return [values, field];
-};
-// TODO CHARLY remove above
 
 export const useSimpleTextValue = ({
     defaultValue,
